Split FileResolver from CloneChallengeCommand
